Rename numpy -> halide_numpy for clarity of include paths
